Move TestFramework Tests to their directory
