+++ analyzers/tests/SonarAnalyzer.UnitTest/TestFramework/Tests/IssueLocationCollectorTest.GetExpectedBuildErrors.cs
@@ -23,10 +23,10 @@ using FluentAssertions;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace SonarAnalyzer.UnitTest.TestFramework.IssueLocationCollectorTests
+namespace SonarAnalyzer.UnitTest.TestFramework.Tests
 {
     [TestClass]
-    public class IssueLocationCollector_GetExpectedBuildErrors
+    public partial class IssueLocationCollectorTest
     {
         [TestMethod]
         public void GetExpectedBuildErrors_No_Comments()
