+++ analyzers/tests/SonarAnalyzer.UnitTest/TestFramework/Tests/IssueLocationCollectorTest.GetExpectedIssueLocations.cs
@@ -24,10 +24,9 @@ using FluentAssertions;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace SonarAnalyzer.UnitTest.TestFramework.IssueLocationCollectorTests
+namespace SonarAnalyzer.UnitTest.TestFramework.Tests
 {
-    [TestClass]
-    public class IssueLocationCollector_GetExpectedIssueLocations
+    public partial class IssueLocationCollectorTest
     {
         [TestMethod]
         public void GetExpectedIssueLocations_No_Comments()
